data step: refactor
Rewrite as a library (set of functions) and a CLI.
